+++ goodtablesio/__init__.py
@@ -1,0 +1,12 @@
+# Register models
+import goodtablesio.models.job
+import goodtablesio.models.user
+
+# Register tasks
+import goodtablesio.tasks.validate
+
+# Register signals
+import goodtablesio.signals
+
+# Register plugins
+import goodtablesio.plugins.github
