+++ goodtablesio/models/__init__.py
@@ -1,3 +1,2 @@
-from . import github_repo
 from . import job
 from . import user
