+++ goodtablesio/plugins/github/__init__.py
@@ -1,0 +1,9 @@
+# Register models
+import goodtablesio.plugins.github.models.repo
+
+# Register tasks
+import goodtablesio.plugins.github.tasks.jobconf
+import goodtablesio.plugins.github.tasks.repos
+
+# Register signals
+import goodtablesio.plugins.github.signals
