+++ goodtablesio/plugins/github/blueprint.py
@@ -2,19 +2,25 @@ import uuid
 import logging
 
 from celery import chain
-from flask import Blueprint, request, abort, jsonify, session, redirect, url_for
+from flask import Blueprint, request, abort, session
+from flask import render_template, jsonify, redirect, url_for
 from flask_login import login_required
 
-from goodtablesio import tasks, models, settings
-
-from goodtablesio.plugins.github.tasks import get_validation_conf, sync_user_repositories
-from goodtablesio.plugins.github.utils import set_commit_status, validate_signature
+from goodtablesio import models, settings
+from goodtablesio.services import database
+from goodtablesio.tasks.validate import validate
+from goodtablesio.plugins.github.models.repo import GithubRepo
+from goodtablesio.plugins.github.tasks.jobconf import get_validation_conf
+from goodtablesio.plugins.github.tasks.repos import sync_user_repositories
+from goodtablesio.plugins.github.utils.status import set_commit_status
+from goodtablesio.plugins.github.utils.signature import validate_signature
+from goodtablesio.plugins.github.utils.hook import get_owner_repo_sha_from_hook_payload
 
 
 log = logging.getLogger(__name__)
 
 
-github = Blueprint('github', __name__, url_prefix='/github')
+github = Blueprint('github', __name__, url_prefix='/github', template_folder='templates')
 
 
 @github.route('/hook', methods=['POST'])
@@ -31,7 +37,7 @@ def create_job():
     payload = request.get_json()
     if not payload:
         abort(400)
-    owner, repo, sha = _get_owner_repo_sha(payload)
+    owner, repo, sha = get_owner_repo_sha_from_hook_payload(payload)
     if not owner:
         abort(400)
 
@@ -61,10 +67,39 @@ def create_job():
     # Run validation
     tasks_chain = chain(
         get_validation_conf.s(owner, repo, sha, job_id=job_id),
-        tasks.validate.s(job_id=job_id))
+        validate.s(job_id=job_id))
     tasks_chain.delay()
 
     return jsonify({'job_id': job_id})
+
+
+@github.route('/repos')
+@login_required
+def repos():
+
+    # Get github syncing status
+    github_sync = False
+    if session.get('github_sync_task_id'):
+        task_id = session['github_sync_task_id']
+        result = sync_user_repositories.AsyncResult(task_id)
+        if result.status == 'PENDING':
+            github_sync = True
+        else:
+            # TODO: cover errors
+            del session['github_sync_task_id']
+
+    # Get github repos
+    github_repos = []
+    if not github_sync:
+        user_id = session['user_id']
+        github_repos = (database['session'].query(GithubRepo).
+            filter(GithubRepo.users.any(id=user_id)).
+            order_by(GithubRepo.owner, GithubRepo.repo).
+            all())
+
+    return render_template('repos.html',
+        github_sync=github_sync,
+        github_repos=github_repos)
 
 
 @github.route('/sync')
@@ -78,30 +113,4 @@ def sync_account():
     # It's kinda general problem it looks like we need
     # to track syncing tasks in the database (github, s3, etc)
     session['github_sync_task_id'] = result.task_id
-    return redirect(url_for('user.projects'))
-
-
-# Internal
-
-def _get_owner_repo_sha(payload):
-
-    try:
-
-        # PR
-        if payload.get('pull_request'):
-            if payload['action'] != 'opened':
-                return None, None, None
-            repo = payload['pull_request']['head']['repo']['name']
-            owner = payload['pull_request']['head']['repo']['owner']['login']
-            sha = payload['pull_request']['head']['sha']
-
-        # PUSH
-        else:
-            repo = payload['repository']['name']
-            owner = payload['repository']['owner']['name']
-            sha = payload['head_commit']['id']
-
-    except KeyError:
-        return None, None, None
-
-    return owner, repo, sha
+    return redirect(url_for('github.repos'))
