+++ goodtablesio/templates/home.html
@@ -3,7 +3,7 @@
 {% block contents %}
 <div class="container">
   <h1>Dashboard</h1>
-  <p>See your <a href="/user/projects">projects</a>.</p>
-  <p>See all <a href="/job">jobs</a>.</p>
+  <p>See your GitHub <a href="/github/repos">repos</a>.</p>
+  <p>See all system <a href="/jobs">jobs</a>.</p>
 </div>
 {% endblock %}
