+++ goodtablesio/tests/plugins/github/tasks/test_jobconf.py
@@ -1,5 +1,5 @@
 from unittest.mock import patch
-from goodtablesio.plugins.github import tasks
+from goodtablesio.plugins.github.tasks import jobconf
 
 # Constants
 
@@ -11,14 +11,14 @@ SHA = 'd5be243487d9882d7f762e7fa04b36b900164a59'
 # Tests
 
 def test_get_job_base():
-    actual = tasks._get_job_base(OWNER, REPO, SHA)
+    actual = jobconf._get_job_base(OWNER, REPO, SHA)
     expect = 'https://raw.githubusercontent.com/frictionlessdata/'
     expect += 'goodtables.io-example/d5be243487d9882d7f762e7fa04b36b900164a59'
     assert actual == expect
 
 
 def test_get_job_files():
-    actual = tasks._get_job_files(OWNER, REPO, SHA)
+    actual = jobconf._get_job_files(OWNER, REPO, SHA)
     expect = [
         'README.md',
         'data/invalid.csv',
@@ -28,10 +28,10 @@ def test_get_job_files():
     assert actual == expect
 
 
-@patch.object(tasks, '_get_job_files_tree_api')
+@patch.object(jobconf, '_get_job_files_tree_api')
 def test_get_job_files_fallback(_get_job_files_tree_api):
     _get_job_files_tree_api.return_value = None
-    actual = tasks._get_job_files(OWNER, REPO, SHA)
+    actual = jobconf._get_job_files(OWNER, REPO, SHA)
     expect = [
         'README.md',
         'data/invalid.csv',
